# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Streaming de Cursos",
        "description": "API para gerenciamento de cursos, usuários e assinaturas. Para usar a API, faça autenticação pelo endpoint /api/login ou usando OAuth2.",
        "contact": {
            "name": "Suporte API",
            "email": "contato@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/admin/users": {
            "get": {
                "tags": [
                    "Administração",
                    "Usuários"
                ],
                "summary": "Listar usuários",
                "description": "Lista todos os usuários do sistema com paginação",
                "operationId": "listUsers",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca para filtrar usuários por nome ou email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Usuário Teste"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "teste@example.com"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "admin"
                                                    },
                                                    "permissions": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido - usuário não tem permissão de admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{user}/role": {
            "put": {
                "tags": [
                    "Administração",
                    "Usuários"
                ],
                "summary": "Atualizar papel do usuário",
                "description": "Atualiza o papel (role) de um usuário específico",
                "operationId": "updateUserRole",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "manager",
                                            "user"
                                        ],
                                        "example": "manager"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Papel atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Papel atualizado com sucesso"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Usuário Teste"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "teste@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "manager"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido - usuário não tem permissão de admin"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{user}/permissions": {
            "put": {
                "tags": [
                    "Administração",
                    "Usuários"
                ],
                "summary": "Atualizar permissões do usuário",
                "description": "Atualiza as permissões de um usuário específico",
                "operationId": "updateUserPermissions",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permissions"
                                ],
                                "properties": {
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "create_content",
                                                "edit_content",
                                                "delete_content",
                                                "manage_users"
                                            ]
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permissões atualizadas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Permissões atualizadas com sucesso"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Usuário Teste"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "teste@example.com"
                                                },
                                                "permissions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido - usuário não tem permissão de admin"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{user}/promote-to-admin": {
            "post": {
                "tags": [
                    "Administração",
                    "Usuários"
                ],
                "summary": "Promover usuário a administrador",
                "description": "Promove um usuário a administrador com todas as permissões",
                "operationId": "promoteUserToAdmin",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário promovido a administrador com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário promovido a administrador com sucesso"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Usuário Teste"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "teste@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                },
                                                "permissions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido - usuário não tem permissão de admin"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "oauth2": {
                "type": "oauth2",
                "description": "OAuth2 para autenticação de API",
                "flows": {
                    "password": {
                        "tokenUrl": "http://localhost/oauth/token",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}